requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-sql-python-connector==1.18.0
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
dnspython==2.7.0
duckdb==1.2.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fonttools==4.57.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
greenlet==3.2.0
h11==0.14.0
html5lib==1.1
idna==3.10
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mplsoccer==1.4.0
multidict==6.4.3
narwhals==1.30.0
nest-asyncio==1.6.0
numpy==2.2.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pg8000==1.31.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
pywin32==309
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scipy==1.15.2
scramp==1.4.5
seaborn==0.13.2
selenium==4.29.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soccerdata==1.8.7
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
stack-data==0.6.3
streamlit==1.43.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
unicode==2.9
Unidecode==1.3.8
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.19.0
